refactor: content storage
move ContentStorage interface to storage
